fix: fixes needed for demo
