+++ flashcard-roguelike/game/entity/dungeon_generator/rooms/event_rooms/ShootingEventRoom.cs
@@ -9,7 +9,7 @@ public partial class ShootingEventRoom : Room, IEventRoom
     [Export] public Turret RoomTurret;
     [Export] public AudioStream CorrectSound;
     [Export] public AudioStream WrongSound;
-    [Export] public TreasureChestOld Reward;
+    [Export] public TreasureChest Reward;
     [Export] public int PenaltyDamage = 25;
 
     public bool IsCompleted { get; private set; }
