--- flashcard-roguelike/game/entity/treasure/TreasureChestOld.cs
@@ -1,248 +1,0 @@
-using Godot;
-using System.Collections.Generic;
-using System.Linq;
-
-public partial class TreasureChestOld : Interactable
-{
-	[Export] public Godot.Collections.Array<ItemResource> LootPool = new();
-	[Export] public PackedScene ItemScene;
-	[Export] public int MinItems        = 1;
-	[Export] public int MaxItems        = 3;
-	[Export] public float SpawnRadius   = 1.5f;
-	[Export] public float SpawnHeight   = 0.5f;
-	[Export] public float SpawnArcStart = 90f;
-	[Export] public float SpawnArcEnd   = 260f;
-	[Export] public bool IsOpen { get; private set; } = false;
-
-	[Signal] public delegate void ChestOpenedEventHandler(TreasureChestOld chest);
-
-	[Export] private Label3D _label;
-	[Export] private RigidBody3D _lidBody;
-
-	[Export] public AudioStream OpenSound;
-	[Export] private AudioStreamPlayer3D _openSoundPlayer;
-
-	private RandomNumberGenerator _rng = new();
-	private List<ItemResource> _rolledItems = new();
-	private int _targetRarityIndex = 0;
-	private Player _playerRef;
-
-	private static readonly Dictionary<int, float> RarityWeights = new()
-	{
-		{ 1, 0.50f },
-		{ 2, 0.25f },
-		{ 3, 0.15f },
-		{ 4, 0.08f },
-		{ 5, 0.02f }
-	};
-
-	public override void _Ready()
-	{
-		base._Ready();
-
-		if (RarityColors.Count() != RarityWeights.Count())
-		{
-			GD.PushError($"The amount of rarities and colors need to match in the treasure room scene: {Name}");
-			return;
-		}
-
-		if (CameraAngles.Count() != RarityWeights.Count())
-		{
-			GD.PushError($"The amount of rarities and camera angles need to match in the treasure room scene: {Name}");
-			return;
-		}
-
-		_rng.Randomize();
-		UpdateLabel();
-		_label.Visible = false;
-
-		if (_openSoundPlayer != null && OpenSound != null)
-		{
-			_openSoundPlayer.Stream = OpenSound;
-		}
-
-		StartIdleShake();
-		StartLightCycle();
-
-		RollLoot();
-
-		StandardMaterial3D particleMat = new();
-		particleMat.AlbedoColor = RarityColors[_targetRarityIndex];
-		particleMat.EmissionEnabled = true;
-		particleMat.Emission = RarityColors[_targetRarityIndex];
-		_particles.DrawPass1.SurfaceSetMaterial(0, particleMat);
-	}
-
-	public override void Interact(Node caller)
-	{
-		OpenChest();
-	}
-
-	public void OpenChest()
-	{
-		if (IsOpen)
-		{
-			return;
-		}
-
-		if (_playerRef == null)
-		{
-			_playerRef = GetTree().GetFirstNodeInGroup("player") as Player;
-			if (_playerRef == null)
-			{
-				GD.PushError("Could not find player.");
-				return;
-			}
-		}
-
-		IsOpen = true;
-		TaloTelemetry.TrackChestsOpened();
-
-		_idleShakeTween?.Kill();
-		_idleShakeTween = null;
-		_lightCycleTween?.Kill();
-		_lightCycleTween = null;
-		_hoverTween?.Kill();
-		_hoverTween = null;
-
-		if (ChestLight != null)
-		{
-			CreateTween().TweenProperty(ChestLight, "light_energy", ChestLight.LightEnergy * 2.5f, 0.3f)
-				.SetTrans(Tween.TransitionType.Quad).SetEase(Tween.EaseType.Out);
-		}
-
-		DoAnticipationShake(FinishOpen);
-	}
-
-	private void FinishOpen()
-	{
-		StartRarityReveal(_targetRarityIndex);
-		LaunchLid();
-
-		if (_particles != null)
-		{
-			_particles.Restart();
-		}
-
-		UpdateLabel();
-		EmitSignal(SignalName.ChestOpened, this);
-	}
-
-	private void SpawnItems()
-	{
-		AudioManager.Instance.PlayGameVictorySound();
-
-		if (ItemScene == null || LootPool.Count == 0)
-		{
-			GD.PushWarning($"TreasureChest: ItemScene={ItemScene != null}, LootPool.Count={LootPool.Count}");
-			return;
-		}
-
-		for (int i = 0; i < _rolledItems.Count; i++)
-		{
-			var entry = _rolledItems[i];
-
-			if (entry == null) { GD.PushWarning("TreasureChest: RollEntry returned null"); continue; }
-
-			var item = ItemScene.Instantiate<Item>();
-			if (item == null) { GD.PushWarning("TreasureChest: Instantiate<Item>() returned null — is the scene root typed as Item?"); continue; }
-
-			item._resource = entry;
-			item.TopLevel = true;
-			GetParent().AddChild(item);
-
-			float t = _rolledItems.Count > 1 ? i / (float)(_rolledItems.Count - 1) : 0.5f;
-			float angle = Mathf.DegToRad(Mathf.Lerp(SpawnArcStart, SpawnArcEnd, t));
-			item.GlobalPosition = GlobalPosition + new Vector3(
-				Mathf.Sin(angle) * SpawnRadius,
-				SpawnHeight,
-				-Mathf.Cos(angle) * SpawnRadius
-			);
-		}
-
-		Timer delay = new();
-		delay.OneShot = true;
-		delay.WaitTime = 2f;
-		delay.Autostart = true;
-
-		delay.Timeout += () =>
-		{
-			_playerRef.PlayerCamera.Current = true;
-		};
-
-		AddChild(delay);
-	}
-
-	private ItemResource RollEntry()
-	{
-		float roll = _rng.Randf();
-		float cumulative = 0f;
-		int rolledRarity = 1;
-
-		foreach (var (rarity, weight) in RarityWeights.OrderByDescending(r => r.Key))
-		{
-			cumulative += weight;
-			if (roll < cumulative) { rolledRarity = rarity; break; }
-		}
-
-		var candidates = LootPool.Where(e => e?.Rarity == rolledRarity).ToList();
-
-		if (candidates.Count == 0)
-		{
-			candidates = LootPool.OrderBy(e => System.Math.Abs(e.Rarity - rolledRarity)).ToList();
-		}
-
-		return candidates.Count > 0 ? candidates[_rng.RandiRange(0, candidates.Count - 1)] : null;
-	}
-
-	private void RollLoot()
-	{
-		_rolledItems.Clear();
-
-		int itemCount = _rng.RandiRange(MinItems, MaxItems);
-
-		for (int i = 0; i < itemCount; i++)
-		{
-			var item = RollEntry();
-			if (item != null)
-			{
-				_rolledItems.Add(item);
-			}
-		}
-
-		_targetRarityIndex = _rolledItems.Count > 0
-			? Mathf.Clamp(_rolledItems.Max(i => i.Rarity) - 1, 0, 4)
-			: 0;
-	}
-
-	private void UpdateLabel()
-	{
-		if (_label == null) return;
-		_label.Text = IsOpen ? "Treasure Chest (Empty)" : "Treasure Chest";
-	}
-
-	public override void HoverStart(Node caller)
-	{
-		_label.Visible = true;
-
-		if (ChestLight == null || IsOpen)
-		{
-			return;
-		}
-
-		_hoverTween?.Kill();
-		float baseEnergy = ChestLight.LightEnergy;
-		_hoverTween = CreateTween().SetLoops();
-		_hoverTween.TweenProperty(ChestLight, "light_energy", baseEnergy * 2f, 0.6f)
-			.SetTrans(Tween.TransitionType.Sine).SetEase(Tween.EaseType.InOut);
-		_hoverTween.TweenProperty(ChestLight, "light_energy", baseEnergy, 0.6f)
-			.SetTrans(Tween.TransitionType.Sine).SetEase(Tween.EaseType.InOut);
-	}
-
-	public override void HoverEnd(Node caller)
-	{
-		_label.Visible = false;
-		_hoverTween?.Kill();
-		_hoverTween = null;
-	}
-}
